Move verification tests into `verification.spec.ts`
